# flockr-admin/src/main/resources/webroot/swagger/flockr-admin.yaml
openapi: 3.0.1
info:
  title: Flockr Admin API
  description: "Flockr Admin Service - Manage audiences, rules, and data connectors"
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Local development server
...
paths:
  /v1/audiences:
    get:
      tags:
      - Audiences
      summary: List audiences with filters
      description: "Retrieves a list of audiences with basic metadata including rule\
        \ counts. Supports filtering by name, creator, and verification status."
      operationId: getAudiencesList
      parameters:
      - name: X-Tenant-Id
        in: header
        description: Tenant identifier
        required: true
        schema:
          type: string
      - name: X-Project-Id
        in: header
        description: Project identifier
        required: true
        schema:
          type: string
      - name: nameSearch
        in: query
        description: Search audiences by name (partial match)
        schema:
          type: string
      - name: createdBy
        in: query
        description: Filter by creator username
        schema:
          type: string
      - name: verified
        in: query
        description: Filter by verification status
        schema:
          type: boolean
      - name: pageSize
        in: query
        description: Number of items per page
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      - name: page
        in: query
        description: Page number (0-indexed)
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      responses:
        "200":
          description: Successfully retrieved audiences list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
    post:
      tags:
      - Audiences
      summary: Create a new audience
      description: Creates a new audience and returns the audience ID
      operationId: createAudience
      parameters:
      - name: X-Tenant-Id
        in: header
        description: Tenant identifier
        required: true
        schema:
          type: string
      - name: X-Project-Id
        in: header
        description: Project identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAudienceRequest'
      responses:
        "200":
          description: Audience created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid/missing body params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/audiences/{audienceId}/rules:
    post:
      tags:
      - Audiences
      summary: Create rules for an audience
      description: Creates rules for a specific audience
      operationId: createRules
      parameters:
      - name: X-Tenant-Id
        in: header
        description: Tenant identifier
        required: true
        schema:
          type: string
      - name: X-Project-Id
        in: header
        description: Project identifier
        required: true
        schema:
          type: string
      - name: audienceId
        in: path
        description: ID of the audience
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRulesRequest'
      responses:
        "200":
          description: Rules created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid/missing body params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/audiences/{audienceId}:
    get:
      tags:
      - Audiences
      summary: Get audience details
      description: Retrieves detailed information about a specific audience
      operationId: getAudienceDetails
      parameters:
      - name: X-Tenant-Id
        in: header
        description: Tenant identifier
        required: true
        schema:
          type: string
      - name: X-Project-Id
        in: header
        description: Project identifier
        required: true
        schema:
          type: string
      - name: audienceId
        in: path
        description: ID of the audience to retrieve
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "304":
          description: Successful Not Modified Response
        "400":
          description: Bad Request due to invalid/missing body params / header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/audiences/{audienceId}/rules/{ruleId}:
    get:
      tags:
      - Audiences
      summary: Get rule details
      description: Retrieves detailed information about a specific rule within an
        audience
      operationId: getRuleDetails
      parameters:
      - name: X-Tenant-Id
        in: header
        description: Tenant identifier
        required: true
        schema:
          type: string
      - name: X-Project-Id
        in: header
        description: Project identifier
        required: true
        schema:
          type: string
      - name: audienceId
        in: path
        description: ID of the audience
        required: true
        schema:
          type: integer
          format: int64
      - name: ruleId
        in: path
        description: ID of the rule to retrieve
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request due to invalid/missing parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "404":
          description: Rule not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
...
components:
  schemas:
    Success:
      type: object
      properties:
        data:
          type: object
    ErrorEntity:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        cause:
          type: string
    Failure:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorEntity'
    CreateAudienceRequest:
      required:
      - description
      - expireDate
      - name
      - sinkIds
      - type
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        customAudienceConfig:
          $ref: '#/components/schemas/JsonObject'
        type:
          type: string
        expireDate:
          maximum: 2177452799000
          type: integer
          format: int64
        sinkIds:
          type: array
          items:
            type: integer
            format: int64
    JsonObject:
      type: object
      properties:
        map:
          type: object
          additionalProperties:
            type: object
        empty:
          type: boolean
...
    CreateRulesRequest:
      required:
      - rules
      type: object
      properties:
        audienceId:
          type: integer
          format: int64
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
...
    Rule:
      required:
      - configuration
      - description
      - end_time
      - name
      - rule_action
      - rule_type
      - start_time
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        configuration:
          $ref: '#/components/schemas/RuleConfigurationSourceInfo'
        start_time:
          type: integer
          format: int64
        end_time:
          type: integer
          format: int64
        rule_type:
          type: string
          enum:
          - STREAM
          - BATCH
        rule_action:
          type: string
          enum:
          - ADD
          - REMOVE
    RuleConfigurationSourceInfo:
      required:
      - configuration_type
      type: object
      properties:
        configuration_type:
          type: string
      discriminator:
        propertyName: configuration_type
...